+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+## [Unreleased]
+### Added
+- Added GPT-4-turbo name variations to cached OpenAI model IDs. This means that we'll
+  be able to see if a model ID should be an OpenAI model, without an OpenAI API key.
+
+
 ## [v12.5.1] - 2024-04-03
 ### Security
 - Now uses an access token to access datasets, allowing the datasets to not be